refactor: move interfaces to its own file
Move the interfaces for bellwether-styles to a type file.

Please note that the themes folder is intentionally left for later
changes in this version.

Issue: #40
