+++ src/index.ts
@@ -1,2 +1,2 @@
 export { flat, material, x11 } from './colors';
-export { Font, Palette, Theme } from './themes';
+export { Font, Palette, Theme } from './types';
