--- src/themes/index.ts
@@ -1,1 +1,0 @@
-export { Font, Palette, Theme } from './interface';
